test/requirements.txt
allure-pytest--2.13.5 llure-python-commons==2.13.5
attrs--23.2.0
certifi--2024.7.4
cffi--1.16.0
charset-normalizer==3.3.2
colorama -0.4.6
distlib=-0.3.8
filelock=-3.16.1
flake8=-7.1.1
greenlet--3.1.0
h11==8.14. idna--3.7
iniconfig--2.0.0 mccabe--0.7.0
outcome--1.3.0.poste
packaging==24.1 pep8--1.7.1
platformdirs==4.3.6 pluggy--1.5.0
psycopg -3.2.2 psycopg-binary--3.2.2 psycopg2-binary--2.9.9 pycodestyle--2.12.1
pycparser--2.22
pyflakes=-3.2.0
PySocks==1.7.1
pytest=-8.2.2
python-dotenv=-1.0.1 requests -2.32.3
selenium--4.23.1
sniffio--1.3.1
sortedcontainers--2.4.0
SQLAlchemy--2.0.35
trio--0.26.0
trio-websocket--0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3-2.2.2
virtualenv=-20.26.5
webdriver-manager--4.0.2
websocket-client==1.8.0
wsproto--1.2.0